requirements.txt
Cartopy>=0.22.0
googlemaps>=4.10.0
matplotlib>=3.8.2
numpy>=1.21.5
pandas>=2.2.1
python-dotenv>=1.0.1
requests_toolbelt>=1.0.0
scipy>=1.12.0
setuptools>=59.6.0
numpy_financial-1.0.0
